rename datadog-remote-config
