+++ datadog-sidecar/Cargo.toml
@@ -27,7 +27,7 @@ libdd-telemetry = { path = "../libdd-telemetry", features = ["tracing"] }
 libdd-data-pipeline = { path = "../libdd-data-pipeline" }
 libdd-trace-utils = { path = "../libdd-trace-utils" }
 libdd-trace-stats = { path = "../libdd-trace-stats", default-features=false, features = ["https"] }
-datadog-remote-config = { path = "../datadog-remote-config" }
+libdd-remote-config = { path = "../libdd-remote-config" }
 datadog-live-debugger = { path = "../datadog-live-debugger" }
 datadog-ffe = { path = "../datadog-ffe", features = ["exposure-events", "evaluation-metrics"] }
 libdd-crashtracker = { path = "../libdd-crashtracker" }
@@ -115,7 +115,7 @@ microseh = "0.1.1"
 libc = { version = "0.2" }
 tempfile = { version = "3.3" }
 httpmock = "0.8.0-alpha.1"
-datadog-remote-config = { path = "../datadog-remote-config", features = ["test"] }
+libdd-remote-config = { path = "../libdd-remote-config", features = ["test"] }
 libdd-trace-utils = { path = "../libdd-trace-utils", features = ["test-utils"] }
 
 [lints.rust]
