+++ datadog-sidecar/src/service/mod.rs
@@ -6,9 +6,9 @@ use crate::config;
 pub use datadog_ffe::telemetry::evaluation_metrics::FfeEvaluationMetric;
 pub use datadog_ffe::telemetry::exposures::{FfeExposure, FfeExposureBatch};
 pub use datadog_ffe::telemetry::FfeTelemetryContext;
-use datadog_remote_config::{RemoteConfigCapabilities, RemoteConfigProduct};
 use libdd_common::tag::Tag;
 use libdd_common::Endpoint;
+use libdd_remote_config::{RemoteConfigCapabilities, RemoteConfigProduct};
 use libdd_telemetry::worker::TelemetryActions;
 use serde::{Deserialize, Serialize};
 use std::path::PathBuf;
