+++ datadog-sidecar/src/service/remote_configs.rs
@@ -3,10 +3,10 @@
 
 use crate::service::{DynamicInstrumentationConfigState, InstanceId};
 use crate::shm_remote_config::{ShmRemoteConfigs, ShmRemoteConfigsGuard};
-use datadog_remote_config::fetch::{
+use libdd_common::{tag::Tag, MutexExt};
+use libdd_remote_config::fetch::{
     ConfigInvariants, ConfigOptions, MultiTargetStats, NotifyTarget, ProductCapabilities,
 };
-use libdd_common::{tag::Tag, MutexExt};
 use std::collections::hash_map::Entry;
 use std::fmt::Debug;
 use std::sync::{Arc, Mutex};
