+++ datadog-sidecar/src/service/session_info.rs
@@ -13,8 +13,8 @@ use futures::future;
 use crate::log::{MultiEnvFilterGuard, MultiWriterGuard};
 use crate::{spawn_map_err, tracer};
 use datadog_live_debugger::sender::{DebuggerType, PayloadSender};
-use datadog_remote_config::fetch::ConfigOptions;
 use libdd_common::{tag::Tag, Endpoint, MutexExt};
+use libdd_remote_config::fetch::ConfigOptions;
 use tracing::{debug, error, info, trace, warn};
 
 use crate::service::agent_info::AgentInfoGuard;
