+++ libdd-remote-config/Cargo.toml
@@ -1,9 +1,13 @@
 [package]
 edition = "2021"
-license = "Apache 2.0"
-name = "datadog-remote-config"
-version = "0.0.1"
-publish = false
+name = "libdd-remote-config"
+version = "0.1.0"
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+homepage = "https://github.com/DataDog/libdatadog/tree/main/libdd-remote-config"
+repository = "https://github.com/DataDog/libdatadog/tree/main/libdd-remote-config"
+description = "Datadog Remote Configuration client and config parsers"
 
 [features]
 default = ["client"]
@@ -23,12 +27,16 @@ client = [
 ]
 
 regex-lite = ["libdd-common/regex-lite"]
+# Enable HTTPS support in the fetcher (rustls + ring via libdd-common).
+https = ["libdd-common/https"]
+# FIPS-compliant crypto provider (aws-lc-rs via libdd-common). Unix only.
+fips = ["libdd-common/fips"]
 test = ["hyper/server", "hyper-util"]
 
 [dependencies]
 anyhow = { version = "1.0" }
-libdd-common = { path = "../libdd-common"}
-libdd-trace-protobuf = { path = "../libdd-trace-protobuf", optional = true }
+libdd-common = { path = "../libdd-common", version = "4.2.0", default-features = false }
+libdd-trace-protobuf = { path = "../libdd-trace-protobuf", version = "3.0.2", optional = true }
 hyper = { workspace = true, optional = true, default-features = false }
 http-body-util = {version = "0.1", optional = true }
 http = { version = "1.1", optional = true }
@@ -51,7 +59,7 @@ hyper-util = { workspace = true, features = ["service"], optional = true }
 
 [dev-dependencies]
 futures = "0.3"
-datadog-remote-config = { path = ".", features = ["test"] }
+libdd-remote-config = { path = ".", features = ["test"] }
 
 [lib]
 bench = false
