+++ libdd-remote-config/examples/remote_config_fetch.rs
@@ -1,13 +1,13 @@
 // Copyright 2021-Present Datadog, Inc. https://www.datadoghq.com/
 // SPDX-License-Identifier: Apache-2.0
 
-use datadog_remote_config::fetch::{ConfigInvariants, ConfigOptions, SingleChangesFetcher};
-use datadog_remote_config::file_change_tracker::{Change, FilePath};
-use datadog_remote_config::file_storage::ParsedFileStorage;
-use datadog_remote_config::RemoteConfigProduct::ApmTracing;
-use datadog_remote_config::{RemoteConfigParsed, Target};
 use libdd_common::tag::Tag;
 use libdd_common::Endpoint;
+use libdd_remote_config::fetch::{ConfigInvariants, ConfigOptions, SingleChangesFetcher};
+use libdd_remote_config::file_change_tracker::{Change, FilePath};
+use libdd_remote_config::file_storage::ParsedFileStorage;
+use libdd_remote_config::RemoteConfigProduct::ApmTracing;
+use libdd_remote_config::{RemoteConfigParsed, Target};
 use std::time::Duration;
 use tokio::time::sleep;
