+++ libdd-remote-config/src/config/agent_task.rs
@@ -72,7 +72,7 @@ pub struct AgentTask {
 /// # Examples
 ///
 /// ```
-/// use datadog_remote_config::config::agent_task::parse_json;
+/// use libdd_remote_config::config::agent_task::parse_json;
 ///
 /// let json_data = r#"{
 ///     "args": {
