+++ libdd-remote-config/src/fetch/shared.rs
