+++ libdd-tracer-flare/Cargo.toml
@@ -13,7 +13,7 @@ repository = "https://github.com/DataDog/libdatadog/tree/main/libdd-tracer-flare
 
 [dependencies]
 anyhow = "1.0"
-datadog-remote-config = { path = "../datadog-remote-config", default-features = false }
+libdd-remote-config = { path = "../libdd-remote-config", default-features = false }
 libdd-common = { version = "4.2.0", path = "../libdd-common" }
 libdd-trace-utils = { version = "6.0.1", path = "../libdd-trace-utils" }
 http = "1"
@@ -28,11 +28,11 @@ tempfile = "3.8"
 bench = false
 
 [dev-dependencies]
-datadog-remote-config = { path = "../datadog-remote-config", features = ["test"] }
+libdd-remote-config = { path = "../libdd-remote-config", features = ["test"] }
 httpmock = "0.8.0-alpha.1"
 libdd-trace-utils = { path = "../libdd-trace-utils", features = ["test-utils"] }
 tokio = { version = "1.36.0", features = ["time", "macros", "rt"] }
 
 [features]
 default = ["listener"]
-listener = ["datadog-remote-config/client"]
+listener = ["libdd-remote-config/client"]
