+++ tools/docker/Dockerfile.build
@@ -96,7 +96,7 @@ COPY "libdd-profiling/Cargo.toml" "libdd-profiling/"
 COPY "libdd-profiling-ffi/Cargo.toml" "libdd-profiling-ffi/"
 COPY "libdd-profiling-protobuf/Cargo.toml" "libdd-profiling-protobuf/"
 COPY "datadog-profiling-replayer/Cargo.toml" "datadog-profiling-replayer/"
-COPY "datadog-remote-config/Cargo.toml" "datadog-remote-config/"
+COPY "libdd-remote-config/Cargo.toml" "libdd-remote-config/"
 COPY "datadog-sidecar/Cargo.toml" "datadog-sidecar/"
 COPY "datadog-sidecar-macros/Cargo.toml" "datadog-sidecar-macros/"
 COPY "datadog-sidecar-ffi/Cargo.toml" "datadog-sidecar-ffi/"
